+++ c-scape/src/fs/lseek.rs
@@ -28,6 +28,7 @@ unsafe extern "C" fn lseek64(fd: c_int, offset: off64_t, whence: c_int) -> off64
         libc::SEEK_CUR => rustix::io::SeekFrom::Current(offset),
         libc::SEEK_END => rustix::io::SeekFrom::End(offset),
         _ => {
+            // TODO: we're missing some stuff here
             set_errno(Errno(libc::EINVAL));
             return -1;
         },
